+++ ext_emconf.php
@@ -9,9 +9,8 @@ $EM_CONF[$_EXTKEY] = [
     'version' => '5.0.0-dev',
     'constraints' => [
         'depends' => [
-            'php' => '8.1.0-0.0.0',
-            'typo3' => '12.4.9-13.4.99',
-            'form' => '12.4.9-13.4.99',
+            'typo3' => '13.4.0-13.4.99',
+            'form' => '13.4.0-13.4.99',
             'jobrouter_base' => '4.0.0-4.99.99',
             'jobrouter_connector' => '4.0.0-4.99.99',
         ],